add json schema to scheduler strategy
